+++ app/src/androidTest/java/woowacourse/movie/movie/java/SettingFragmentTest.kt
@@ -15,7 +15,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import woowacourse.movie.R
 import woowacourse.movie.movie.SettingPreference
-import woowacourse.movie.movielist.MainActivity
+import woowacourse.movie.movie.activity.MainActivity
 
 @RunWith(AndroidJUnit4::class)
 class SettingFragmentTest {
