+++ app/src/main/java/woowacourse/movie/movie/dto/movie/MovieTimeDto.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.dto
+package woowacourse.movie.movie.dto.movie
 
 import java.io.Serializable
 import java.time.LocalTime
