+++ app/src/main/java/woowacourse/movie/movie/dto/seat/SeatColDto.kt
@@ -1,3 +1,3 @@
-package woowacourse.movie.movie.dto
+package woowacourse.movie.movie.dto.seat
 
 data class SeatColDto(val col: Int) : java.io.Serializable
