+++ app/src/main/java/woowacourse/movie/movie/dto/ticket/TicketCountDto.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.dto
+package woowacourse.movie.movie.dto.ticket
 
 import java.io.Serializable
