+++ app/src/main/java/woowacourse/movie/movie/fragment/HomeFragment.kt
@@ -10,11 +10,11 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
 import woowacourse.movie.dto.AdDto
-import woowacourse.movie.dto.MovieDto
-import woowacourse.movie.dto.MovieDummy
-import woowacourse.movie.moviedetail.MovieDetailActivity
-import woowacourse.movie.movielist.MovieRVAdapter
-import woowacourse.movie.movielist.OnClickListener
+import woowacourse.movie.movie.activity.MovieDetailActivity
+import woowacourse.movie.movie.dto.movie.MovieDto
+import woowacourse.movie.movie.dto.movie.MovieDummy
+import woowacourse.movie.movie.movielist.MovieRVAdapter
+import woowacourse.movie.movie.movielist.OnClickListener
 
 class HomeFragment : Fragment() {
