+++ app/src/main/java/woowacourse/movie/movie/movielist/AdViewHolder.kt
@@ -1,10 +1,11 @@
-package woowacourse.movie.movielist
+package woowacourse.movie.movie.movielist
 
 import android.view.View
 import android.widget.ImageView
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
 import woowacourse.movie.dto.AdDto
+import woowacourse.movie.movie.movielist.OnClickListener
 
 class AdViewHolder(private val view: View, private val onAdClickListener: OnClickListener<AdDto>) :
     RecyclerView.ViewHolder(view) {
