+++ app/src/main/java/woowacourse/movie/movie/movielist/OnClickListener.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.movielist
+package woowacourse.movie.movie.movielist
 
 interface OnClickListener<T> {
