+++ app/src/main/java/woowacourse/movie/movie/movielist/ViewType.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.movielist
+package woowacourse.movie.movie.movielist
 
 enum class ViewType {
     MOVIE_VIEW, AD_VIEW
